+++ grid_map_filters/test/empty_test.cpp
@@ -1,5 +1,0 @@
-#include <gtest/gtest.h>
-
-TEST(Placeholder, emptyTest) {  // NOLINT
-  ASSERT_TRUE(true);
-}